+++ .travis.yml
@@ -36,15 +36,14 @@ jobs:
   include:
     - stage: deploy test server
       env:
-        - BRANCH_LINUX_CHROME=linux-firefox
+        - BRANCH_LINUX_FIREFOX=linux-firefox
       script:
-        - git branch linux-firefox
-        - git push -f origin linux-firefox
+        - git branch ${BRANCH_LINUX_FIREFOX}
+        - git push -f origin ${BRANCH_LINUX_FIREFOX}
     - stage: deploy test server
       env:
         - BRANCH_LINUX_CHROME=linux-chrome
       script:
-        - git pull -f origin master
         - git branch ${BRANCH_LINUX_CHROME}
         - git push -f origin ${BRANCH_LINUX_CHROME}
     - stage: unit test
