+++ SUITES.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 <suite name="UI tests">
     <suite-files>
-        <suite-file path="qa/testLinuxChrome.xml"/>
-        <suite-file path="qa/testLinuxFirefox.xml"/>
+        <suite-file path="testLinuxChrome.xml"/>
+        <suite-file path="testLinuxFirefox.xml"/>
     </suite-files>
 </suite>
