--- qa/src/test/java/uitest/contacts/ContactsListPageTest.java
@@ -1,7 +1,0 @@
-package uitest.contacts;
-
-import uitest.config.BaseTest;
-
-public class ContactsListPageTest extends BaseTest {
-
-}
