--- qa/src/test/java/uitest/organizations/OrganizationsListPageTest.java
@@ -1,7 +1,0 @@
-package uitest.organizations;
-
-import uitest.config.BaseTest;
-
-public class OrganizationsListPageTest extends BaseTest {
-
-}
