+++ src/main/java/selenium/config/BaseTest.java
@@ -1,4 +1,4 @@
-package uitest.config;
+package selenium.config;
 
 import org.openqa.selenium.WebDriver;
