+++ src/main/java/selenium/contacts/ContactsListPage.java
@@ -1,9 +1,10 @@
-package uitest.contacts;
+package selenium.contacts;
 
 import com.codeborne.selenide.SelenideElement;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import uitest.config.BasePage;
+import selenium.config.BasePage;
+
 import static com.codeborne.selenide.Selenide.$;
 
 public class ContactsListPage extends BasePage {
