+++ src/main/java/selenium/organizations/AddOrganizationsPage.java
@@ -1,20 +1,19 @@
-package uitest.contacts;
+package selenium.organizations;
 
 import com.codeborne.selenide.SelenideElement;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import uitest.config.BasePage;
+import selenium.config.BasePage;
+
 import static com.codeborne.selenide.Selenide.$;
 
-public class AddContactsPage extends BasePage {
+public class AddOrganizationsPage extends BasePage {
 
-    public String confirmSuccesfully = "Contact succesfully created!";
+    public String confirmSuccesfully = "Organization succesfully created!";
 
-    SelenideElement fieldFirstName = $(By.name("firstName"));
-    SelenideElement fieldLastName = $(By.name("lastName"));
     SelenideElement fieldEmail = $(By.name("email"));
     SelenideElement fieldPhone = $(By.name("phone"));
-    SelenideElement dropdownOrganization = $(By.name("organizationId"));
+    SelenideElement dropdownOrganizationType = $(By.name("organizationTypeId"));
     SelenideElement fieldStreet = $(By.name("street"));
     SelenideElement fieldCity = $(By.name("city"));
     SelenideElement fieldCountry = $(By.name("country"));
@@ -22,8 +21,9 @@ public class AddContactsPage extends BasePage {
     SelenideElement fieldNumber = $(By.name("number"));
     SelenideElement fieldProvince = $(By.name("province"));
     SelenideElement btnAdd = $(By.xpath("//*[@type='submit']"));
+    SelenideElement fieldOrganizationName = $(By.name("name"));
 
-    public AddContactsPage(WebDriver driver) {
+    public AddOrganizationsPage(WebDriver driver) {
         super(driver);
     }
 }
