+++ src/test/java/selenium/contacts/ContactsListPageTest.java
@@ -1,0 +1,7 @@
+package selenium.contacts;
+
+import selenium.config.BaseTest;
+
+public class ContactsListPageTest extends BaseTest {
+
+}
