+++ src/test/java/selenium/organizations/OrganizationsListPageTest.java
@@ -1,0 +1,7 @@
+package selenium.organizations;
+
+import selenium.config.BaseTest;
+
+public class OrganizationsListPageTest extends BaseTest {
+
+}
