+++ testOsxChrome.xml
@@ -9,9 +9,9 @@
         <parameter name="timeout" value="8000"/>
         <parameter name="baseUrl" value="https://task-delivery-test-osx-chrome.herokuapp.com/"/>
         <classes>
-            <class name="uitest.mainPage.MainPageTest"/>
-            <class name="uitest.contacts.AddContactsPageTest"/>
-            <class name="uitest.organizations.AddOrganizationsPageTest"/>
+            <class name="selenium.mainPage.MainPageTest"/>
+            <class name="selenium.contacts.AddContactsPageTest"/>
+            <class name="selenium.organizations.AddOrganizationsPageTest"/>
         </classes>
     </test>
 </suite>
